pyproject.toml
[project]
name = "google-cloud-mcp"
version = "0.1.0"
description = "MCP server for Google Calendar integration"
requires-python = ">=3.12"
dependencies = [
    "mcp",
    "google-api-python-client",
    "google-auth-oauthlib",
    "google-auth-httplib2",
]

[project.scripts]
google-cloud-mcp = "google_cloud_mcp.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/google_cloud_mcp"]
